+++ eventex/subscriptions/forms.py
@@ -1,14 +1,7 @@
 from django import forms
 from django.core.exceptions import ValidationError
 from eventex.subscriptions.models import Subscription
-
-
-def validate_cpf(value):
-    if not value.isdigit():
-        raise ValidationError('CPF deve conter apenas números.', 'digits')
-
-    if len(value) != 11:
-        raise ValidationError('CPF deve ter 11 números.', 'length')
+from eventex.subscriptions.validators import validate_cpf
 
 
 class SubscriptionFormOld(forms.Form):
