+++ eventex/subscriptions/models.py
@@ -1,9 +1,10 @@
 from django.db import models
+from eventex.subscriptions.validators import validate_cpf
 
 
 class Subscription(models.Model):
 	name = models.CharField('nome', max_length=100)
-	cpf = models.CharField('CPF', max_length=11)
+	cpf = models.CharField('CPF', max_length=11, validators=[validate_cpf])
 	email = models.EmailField('e-mail')
 	phone = models.CharField('telefone', max_length=20)
 	created_at = models.DateTimeField('criado em', auto_now_add=True)
