+++ eventex/subscriptions/validators.py
@@ -1,0 +1,9 @@
+from django.core.exceptions import ValidationError
+
+
+def validate_cpf(value):
+    if not value.isdigit():
+        raise ValidationError('CPF deve conter apenas números.', 'digits')
+
+    if len(value) != 11:
+        raise ValidationError('CPF deve ter 11 números.', 'length')
